add update check command
