+++ cmd/root.go
@@ -1,16 +1,12 @@
 package cmd
 
 import (
-	"context"
 	"fmt"
 	"log"
 	"os"
-	"runtime"
 	"strconv"
-	"time"
 
 	"github.com/byxorna/nycmesh-tool/pkg/app"
-	"github.com/byxorna/nycmesh-tool/pkg/selfupdate"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -36,49 +32,11 @@ func Execute() {
 	}
 }
 
-func checkForUpdate() error {
-	if c, err := selfupdate.NewUpdaterGithub(); err == nil {
-		log.Printf("checking for updates")
-		ctx, cancelFunc := context.WithTimeout(context.Background(), 3*time.Second)
-		defer cancelFunc()
-
-		releaseAssetName := fmt.Sprintf("%s-%s-%s", rootCmd.Use, runtime.GOOS, runtime.GOARCH)
-		if runtime.GOOS == "windows" {
-			releaseAssetName = fmt.Sprintf("%s.exe", releaseAssetName)
-		}
-		r, err := c.HasNewerRelease(ctx, releaseAssetName)
-
-		switch err {
-		case selfupdate.ErrConnectionError:
-			log.Printf("update check failed: %s", err.Error())
-		case selfupdate.ErrNoNewRelease:
-			log.Printf("up to date! (%s)", r.TagName)
-		case nil:
-			if r.DownloadURL != "" {
-				binName := fmt.Sprintf(`$HOME/bin/%s`, releaseAssetName)
-				log.Printf(`🎉 %s is available! Download it with 'curl -o %s "%s" && chmod +x %s'`, binName, r.TagName, r.DownloadURL, binName)
-			} else {
-				// probably, we dont have builds available for this releaseAssetName
-				log.Printf("🥳 %s is available, but I didnt find a release asset named '%s'. See %s for available builds", releaseAssetName, r.TagName, r.URL)
-			}
-		default:
-			log.Printf("update error: %s", err.Error())
-		}
-	}
-	return nil
-}
-
 func init() {
 	cobra.OnInitialize(initConfig)
 	rootCmd.PersistentFlags().StringP("format", "f", "json", "output format (table or json)")
 	viper.BindPFlag("core.format", rootCmd.PersistentFlags().Lookup("format"))
 	rootCmd.PersistentFlags().StringVar(&cfgFile, "config", "", "config file (default is $HOME/.nycmesh-tool.yaml)")
-
-	rootCmd.PersistentPostRun = func(cmd *cobra.Command, args []string) {
-		// if execution was successful, lets try to check for an update
-		// TODO: dont exhaust github rate limit by checking each invocation
-		_ = checkForUpdate()
-	}
 }
 
 // initConfig reads in config file and ENV variables if set.
